+++ extras/tests/JsonArray/copyArray.cpp
@@ -19,6 +19,18 @@ TEST_CASE("copyArray()") {
     REQUIRE(std::string("[1,2,3]") == json);
   }
 
+  SECTION("1D -> JsonDocument") {
+    DynamicJsonDocument doc(4096);
+    char json[32];
+    int source[] = {1, 2, 3};
+
+    bool ok = copyArray(source, doc);
+    REQUIRE(ok);
+
+    serializeJson(doc, json, sizeof(json));
+    REQUIRE(std::string("[1,2,3]") == json);
+  }
+
   SECTION("1D -> JsonArray, but not enough memory") {
     const size_t SIZE = JSON_ARRAY_SIZE(2);
     StaticJsonDocument<SIZE> doc;
@@ -43,6 +55,18 @@ TEST_CASE("copyArray()") {
     REQUIRE(ok);
 
     serializeJson(array, json, sizeof(json));
+    REQUIRE(std::string("[[1,2,3],[4,5,6]]") == json);
+  }
+
+  SECTION("2D -> JsonDocument") {
+    DynamicJsonDocument doc(4096);
+    char json[32];
+    int source[][3] = {{1, 2, 3}, {4, 5, 6}};
+
+    bool ok = copyArray(source, doc);
+    REQUIRE(ok);
+
+    serializeJson(doc, json, sizeof(json));
     REQUIRE(std::string("[[1,2,3],[4,5,6]]") == json);
   }
 
@@ -96,6 +120,22 @@ TEST_CASE("copyArray()") {
     REQUIRE(2 == destination[1]);
   }
 
+  SECTION("JsonDocument -> 1D") {
+    DynamicJsonDocument doc(4096);
+    char json[] = "[1,2,3]";
+    DeserializationError err = deserializeJson(doc, json);
+    REQUIRE(err == DeserializationError::Ok);
+
+    int destination[4] = {0};
+    size_t result = copyArray(doc, destination);
+
+    REQUIRE(3 == result);
+    REQUIRE(1 == destination[0]);
+    REQUIRE(2 == destination[1]);
+    REQUIRE(3 == destination[2]);
+    REQUIRE(0 == destination[3]);
+  }
+
   SECTION("JsonArray -> 2D") {
     DynamicJsonDocument doc(4096);
     char json[] = "[[1,2],[3],[4]]";
@@ -106,6 +146,24 @@ TEST_CASE("copyArray()") {
 
     int destination[3][2] = {{0}};
     copyArray(array, destination);
+
+    REQUIRE(1 == destination[0][0]);
+    REQUIRE(2 == destination[0][1]);
+    REQUIRE(3 == destination[1][0]);
+    REQUIRE(0 == destination[1][1]);
+    REQUIRE(4 == destination[2][0]);
+    REQUIRE(0 == destination[2][1]);
+  }
+
+  SECTION("JsonDocument -> 2D") {
+    DynamicJsonDocument doc(4096);
+    char json[] = "[[1,2],[3],[4]]";
+
+    DeserializationError err = deserializeJson(doc, json);
+    REQUIRE(err == DeserializationError::Ok);
+
+    int destination[3][2] = {{0}};
+    copyArray(doc, destination);
 
     REQUIRE(1 == destination[0][0]);
     REQUIRE(2 == destination[0][1]);