+++ CHANGELOG.md
@@ -7,6 +7,7 @@ HEAD
 * CMake: don't build tests when imported in another project
 * CMake: made project arch-independent
 * Visual Studio: fixed error C2766 with flag `/Zc:__cplusplus` (issue #1250)
+* Added support for `JsonDocument` to `copyArray()` (issue #1255)
 
 v6.15.1 (2020-04-08)
 -------